requirements.txt

smartapi-python
logzero
websocket-client
pyotp
pandas
pytz
matplotlib
seaborn
tqdm
tenacity
python-dotenv
